Minor cleanups.

+++ XTulator/modules/io/bswap.h
@@ -160,8 +160,8 @@ void cpu_to_le16wu(uint16_t* p, uint16_t v)
 {
 	uint8_t* p1 = (uint8_t*)p;
 
-	p1[0] = v;
-	p1[1] = v >> 8;
+	p1[0] = (uint8_t)v;
+	p1[1] = (uint8_t)(v >> 8);
 }
 
 void cpu_to_le32wu(uint32_t* p, uint32_t v)
@@ -190,8 +190,8 @@ void cpu_to_be16wu(uint16_t* p, uint16_t v)
 {
 	uint8_t* p1 = (uint8_t*)p;
 
-	p1[0] = v >> 8;
-	p1[1] = v;
+	p1[0] = (uint8_t)(v >> 8);
+	p1[1] = (uint8_t)v;
 }
 
 void cpu_to_be32wu(uint32_t* p, uint32_t v)
